some refactoring around macros
